iter_to_container -> keys_cache and liberated dropbox from requests: Now a standard lib version using urllib!
